+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Optimize git operations to reduce lock contention.
 * Fix kubernetes discovery module to not override user defined tasks.
 
 # 0.25.1