+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.26.0
+
 * Record GitHub CheckRuns and display them like statuses.
 * Optimize git operations to reduce lock contention.
 * Fix kubernetes discovery module to not override user defined tasks.